pyproject.toml
[project]
name = "veinpay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pymongo>=4.15.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "scikit-image>=0.25.2",
    "tensorflow>=2.20.0",
    "uvicorn>=0.38.0",
]
